+++ tower-buffer/src/error.rs
@@ -14,12 +14,6 @@ pub struct Closed {
     _p: (),
 }
 
-/// Error produced when spawning the worker fails
-#[derive(Debug)]
-pub struct SpawnError {
-    _p: (),
-}
-
 /// Errors produced by `Buffer`.
 pub(crate) type Error = Box<dyn std::error::Error + Send + Sync>;
 
@@ -66,19 +60,3 @@ impl fmt::Display for Closed {
 }
 
 impl std::error::Error for Closed {}
-
-// ===== impl SpawnError =====
-
-impl SpawnError {
-    pub(crate) fn new() -> SpawnError {
-        SpawnError { _p: () }
-    }
-}
-
-impl fmt::Display for SpawnError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "failed to spawn Buffer worker task")
-    }
-}
-
-impl std::error::Error for SpawnError {}