+++ tower-buffer/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/tower-buffer/0.3.0-alpha.2")]
+#![doc(html_root_url = "https://docs.rs/tower-buffer/0.3.0")]
 #![warn(
     missing_debug_implementations,
     missing_docs,
@@ -23,4 +23,3 @@ mod worker;
 
 pub use crate::layer::BufferLayer;
 pub use crate::service::Buffer;
-pub use crate::worker::WorkerExecutor;